refactor(geocode): merge location parsing into one module with a single seam
Candidate 5 from architecture review: location parsing was split across
locationParser.js (raw CLI words → string) and openmeteo.js
(parseLocationQuery + geocode). Two parsers, two seams, one concept.
Understanding 'San Ramon CA' → coordinates required bouncing between both
files plus the geocoding API call.

Extracted parseLocationQuery and geocode into src/api/geocode.js — one
module owns location parsing, API call, and result selection. openmeteo.js
re-exports geocode and parseLocationQuery for backward compatibility. Removed
unused imports (WeatherError, sanitizeForDisplay) from openmeteo.js.

All 629 tests pass. Lint clean.
